filter: Use struct uftrace_filter for match result
I don't see why it needs to copy a trigger as a result.  The only case
was to adjust flags for cygprof and xray to disable arguments and recover
actions.  But this can be handled without changing the flags.

To ensure it's just using the filter entry, let's make the result type
of uftrace_match_filter() const and remove the trigger output argument.

Unfortunately it has to convert a lot of callsites.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>